Extract DuckRel into dedicated module
